Results from a late-night auto mode testing session
I had issues with the insertion of a parallel command group
with the intake, where it would crash because you can't
use the same subsystem in a parallel command group.

I also observed the wrist not working and tried to diagnose,
but later discovered where I believe the wrist is firing
the wrong direction on the solenoid.

I adjusted the PID values to get better alignment on the path

I added a path for scoring 2 cones on grid 2 in auto, seems
to work well.